♻️ refactor: split conditionalSection into sibling extension (#144)
The step-2 additions (`.contextMenu` + discoverability footer) pushed
`PhaseEditorSheet.swift` to 406 lines — over SwiftLint's
`file_length` warning threshold (400). Moving `conditionalSection`
and its `branchSection` helper into a sibling extension file brings
the main file back under 400 (323 lines) and keeps the split roughly
along a natural seam (conditional-specific UI vs. the common
per-phase-type sections).

No behavior change. Widens `SubPhaseEditContext` and `editingSubPhase`
from file-private to module-internal so the sibling extension can
drive nested-sheet presentation.
