+++ Pastura/Pastura/Views/Editor/PhaseEditorSheet.swift
@@ -3,7 +3,11 @@ import SwiftUI
 /// Identifies which branch's sub-phase is currently being edited via a
 /// nested `PhaseEditorSheet`. `.sheet(item:)` drives the presentation;
 /// `onSave` on the nested sheet writes back to the right branch.
-private struct SubPhaseEditContext: Identifiable {
+///
+/// Module-internal (rather than file-private) because the conditional-
+/// section UI lives in a sibling extension file and needs to construct
+/// these contexts.
+struct SubPhaseEditContext: Identifiable {
   let id = UUID()
   let branch: EditablePhase.Branch
   var phase: EditablePhase
@@ -29,7 +33,9 @@ struct PhaseEditorSheet: View {
 
   @State private var newOutputFieldName: String = ""
   @State private var newOptionText: String = ""
-  @State private var editingSubPhase: SubPhaseEditContext?
+  // Internal (not private) so the sibling conditional-section extension
+  // can present the nested editor from the "Add sub-phase" button.
+  @State var editingSubPhase: SubPhaseEditContext?
 
   var body: some View {
     NavigationStack {
@@ -177,98 +183,9 @@ struct PhaseEditorSheet: View {
     }
   }
 
-  private var conditionalSection: some View {
-    Group {
-      Section {
-        TextField(
-          "e.g. max_score >= 10",
-          text: $phase.condition,
-          axis: .vertical
-        )
-        .font(.body.monospaced())
-        .textInputAutocapitalization(.never)
-        .autocorrectionDisabled()
-      } header: {
-        Text("Condition")
-      } footer: {
-        Text(
-          "Single comparison: lhs OP rhs where OP is one of ==, !=, <, <=, >, >=. "
-            + "Identifiers: current_round, total_rounds, max_score, min_score, "
-            + "eliminated_count, active_count, vote_winner, scores.<Name>, "
-            + "or any template variable. Wrap string literals in double quotes."
-        )
-        .font(.caption)
-      }
-
-      branchSection(
-        title: "Then branch (condition true)",
-        phases: $phase.thenPhases,
-        branch: .then
-      )
-      branchSection(
-        title: "Else branch (condition false)",
-        phases: $phase.elsePhases,
-        branch: .else
-      )
-    }
-  }
-
-  @ViewBuilder
-  private func branchSection(
-    title: String,
-    phases: Binding<[EditablePhase]>,
-    branch: EditablePhase.Branch
-  ) -> some View {
-    Section {
-      ForEach(phases.wrappedValue) { subPhase in
-        Button {
-          editingSubPhase = SubPhaseEditContext(branch: branch, phase: subPhase)
-        } label: {
-          HStack {
-            Text(subPhase.type.rawValue)
-              .font(.body.monospaced())
-              .foregroundStyle(.primary)
-            Spacer()
-            Image(systemName: "chevron.right")
-              .foregroundStyle(.secondary)
-              .font(.caption)
-          }
-        }
-        .contextMenu {
-          Button {
-            let other: EditablePhase.Branch = (branch == .then ? .else : .then)
-            phase.moveSubPhase(id: subPhase.id, to: other)
-          } label: {
-            Label(
-              branch == .then ? "Move to Else Branch" : "Move to Then Branch",
-              systemImage: "arrow.left.arrow.right"
-            )
-          }
-        }
-      }
-      .onDelete { indexSet in
-        phases.wrappedValue.remove(atOffsets: indexSet)
-      }
-      .onMove { source, destination in
-        // Within-branch reorder via .onMove. Cross-branch move is exposed
-        // via the .contextMenu on each row.
-        phases.wrappedValue.move(fromOffsets: source, toOffset: destination)
-      }
-
-      Button {
-        let newPhase = EditablePhase()
-        phases.wrappedValue.append(newPhase)
-        editingSubPhase = SubPhaseEditContext(branch: branch, phase: newPhase)
-      } label: {
-        Label("Add sub-phase", systemImage: "plus.circle")
-      }
-    } header: {
-      Text(title)
-    } footer: {
-      Text("Long-press a sub-phase to move it to the other branch.")
-        .font(.caption)
-    }
-  }
+  // `conditionalSection` and its `branchSection` helper live in
+  // `PhaseEditorSheet+ConditionalSection.swift` (sibling extension file)
+  // to keep this file under SwiftLint's `file_length` limit.
 
   // MARK: - Type-Specific Sections
